+++ src/main/java/com/microsoft/graph/models/ManagedDevice.java
@@ -43,7 +43,7 @@ public class ManagedDevice extends Entity implements IJsonBackedObject {
 
     /**
      * The Activation Lock Bypass Code.
-     * Code that allows the Activation Lock on a device to be bypassed. This property is read-only.
+     * The code that allows the Activation Lock on managed device to be bypassed. Default, is Null (Non-Default property) for this property when returned as part of managedDevice entity in LIST call. Individual GET call with select query options is needed to retrieve actual values. Supports: $select. $Search is not supported. Read-only. This property is read-only.
      */
     @SerializedName(value = "activationLockBypassCode", alternate = {"ActivationLockBypassCode"})
     @Expose
@@ -205,7 +205,7 @@ public class ManagedDevice extends Entity implements IJsonBackedObject {
 
     /**
      * The Ethernet Mac Address.
-     * Ethernet MAC. This property is read-only.
+     * Ethernet MAC. Default, is Null (Non-Default property) for this property when returned as part of managedDevice entity. Individual get call with select query options is needed to retrieve actual values. Example: deviceManagement/managedDevices({managedDeviceId})?$select=ethernetMacAddress Supports: $select. $Search is not supported. Read-only. This property is read-only.
      */
     @SerializedName(value = "ethernetMacAddress", alternate = {"EthernetMacAddress"})
     @Expose