+++ src/main/java/com/microsoft/graph/models/NotificationMessageTemplate.java
@@ -30,7 +30,7 @@ public class NotificationMessageTemplate extends Entity implements IJsonBackedOb
 
     /**
      * The Branding Options.
-     * The Message Template Branding Options. Branding is defined in the Intune Admin Console. Possible values are: none, includeCompanyLogo, includeCompanyName, includeContactInformation, includeDeviceDetails.
+     * The Message Template Branding Options. Branding is defined in the Intune Admin Console. Possible values are: none, includeCompanyLogo, includeCompanyName, includeContactInformation, includeCompanyPortalLink, includeDeviceDetails.
      */
     @SerializedName(value = "brandingOptions", alternate = {"BrandingOptions"})
     @Expose