+++ src/main/java/com/microsoft/graph/models/ReportRoot.java
@@ -10,7 +10,6 @@ import com.microsoft.graph.serializer.AdditionalDataManager;
 import java.util.EnumSet;
 import com.microsoft.graph.http.BaseCollectionPage;
 import com.microsoft.graph.models.SecurityReportsRoot;
-import com.microsoft.graph.models.Entity;
 
 
 import com.google.gson.JsonObject;
@@ -24,8 +23,21 @@ import javax.annotation.Nonnull;
 /**
  * The class for the Report Root.
  */
-public class ReportRoot extends Entity implements IJsonBackedObject {
+public class ReportRoot implements IJsonBackedObject {
 
+    /** the OData type of the object as returned by the service */
+    @SerializedName("@odata.type")
+    @Expose
+    @Nullable
+    public String oDataType;
+
+    private transient AdditionalDataManager additionalDataManager = new AdditionalDataManager(this);
+
+    @Override
+    @Nonnull
+    public final AdditionalDataManager additionalDataManager() {
+        return additionalDataManager;
+    }
 
     /**
      * The Daily Print Usage By Printer.